requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=15.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict<=0.10.0
torchdata
transformers
# vllm==0.6.3.post1
wandb
packaging>=20.0
uvicorn
fastapi
